# Optimize storage for segment information