--- a/src/Dependencies/Collections/Internal/SegmentedArrayHelper.cs
+++ b/src/Dependencies/Collections/Internal/SegmentedArrayHelper.cs
@@ -4,6 +4,8 @@
 
 using System;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
 
 namespace Microsoft.CodeAnalysis.Collections.Internal
 {
@@ -15,6 +17,52 @@ namespace Microsoft.CodeAnalysis.Collections.Internal
         internal const int IntrosortSizeThreshold = 16;
 
         /// <summary>
+        /// A combination of <see cref="MethodImplOptions.AggressiveInlining"/> and
+        /// <see cref="F:System.Runtime.CompilerServices.MethodImplOptions.AggressiveOptimization"/>.
+        /// </summary>
+        [SuppressMessage("Documentation", "CA1200:Avoid using cref tags with a prefix", Justification = "The field is not supported in all compilation targets.")]
+        internal const MethodImplOptions FastPathMethodImplOptions = MethodImplOptions.AggressiveInlining | (MethodImplOptions)512;
+
+        [MethodImpl(FastPathMethodImplOptions)]
+        internal static int GetSegmentSize<T>()
+        {
+            if (Unsafe.SizeOf<T>() == Unsafe.SizeOf<object>())
+            {
+                return ReferenceTypeSegmentHelper.SegmentSize;
+            }
+            else
+            {
+                return ValueTypeSegmentHelper<T>.SegmentSize;
+            }
+        }
+
+        [MethodImpl(FastPathMethodImplOptions)]
+        internal static int GetSegmentShift<T>()
+        {
+            if (Unsafe.SizeOf<T>() == Unsafe.SizeOf<object>())
+            {
+                return ReferenceTypeSegmentHelper.SegmentShift;
+            }
+            else
+            {
+                return ValueTypeSegmentHelper<T>.SegmentShift;
+            }
+        }
+
+        [MethodImpl(FastPathMethodImplOptions)]
+        internal static int GetOffsetMask<T>()
+        {
+            if (Unsafe.SizeOf<T>() == Unsafe.SizeOf<object>())
+            {
+                return ReferenceTypeSegmentHelper.OffsetMask;
+            }
+            else
+            {
+                return ValueTypeSegmentHelper<T>.OffsetMask;
+            }
+        }
+
+        /// <summary>
         /// Calculates the maximum number of elements of size <paramref name="elementSize"/> which can fit into an array
         /// which has the following characteristics:
         /// <list type="bullet">
@@ -24,7 +72,7 @@ namespace Microsoft.CodeAnalysis.Collections.Internal
         /// </summary>
         /// <param name="elementSize">The size of the elements in the array.</param>
         /// <returns>The segment size to use for small object heap segmented arrays.</returns>
-        internal static int CalculateSegmentSize(int elementSize)
+        private static int CalculateSegmentSize(int elementSize)
         {
             // Default Large Object Heap size threshold
             // https://github.com/dotnet/runtime/blob/c9d69e38d0e54bea5d188593ef6c3b30139f3ab1/src/coreclr/src/gc/gc.h#L111
@@ -50,7 +98,7 @@ namespace Microsoft.CodeAnalysis.Collections.Internal
         /// </summary>
         /// <param name="segmentSize">The number of elements in each page of the segmented array. Must be a power of 2.</param>
         /// <returns>The shift to apply to the absolute index to get the page index within a segmented array.</returns>
-        internal static int CalculateSegmentShift(int segmentSize)
+        private static int CalculateSegmentShift(int segmentSize)
         {
             var segmentShift = 0;
             while (0 != (segmentSize >>= 1))
@@ -67,10 +115,36 @@ namespace Microsoft.CodeAnalysis.Collections.Internal
         /// </summary>
         /// <param name="segmentSize">The number of elements in each page of the segmented array. Must be a power of 2.</param>
         /// <returns>The bit mask to obtain the index within a page from an absolute index within a segmented array.</returns>
-        internal static int CalculateOffsetMask(int segmentSize)
+        private static int CalculateOffsetMask(int segmentSize)
         {
             Debug.Assert(segmentSize == 1 || (segmentSize & (segmentSize - 1)) == 0, "Expected size of 1, or a power of 2");
             return segmentSize - 1;
+        }
+
+        internal static class TestAccessor
+        {
+            public static int CalculateSegmentSize(int elementSize)
+                => SegmentedArrayHelper.CalculateSegmentSize(elementSize);
+
+            public static int CalculateSegmentShift(int elementSize)
+                => SegmentedArrayHelper.CalculateSegmentShift(elementSize);
+
+            public static int CalculateOffsetMask(int elementSize)
+                => SegmentedArrayHelper.CalculateOffsetMask(elementSize);
+        }
+
+        private static class ReferenceTypeSegmentHelper
+        {
+            public static readonly int SegmentSize = CalculateSegmentSize(Unsafe.SizeOf<object>());
+            public static readonly int SegmentShift = CalculateSegmentShift(SegmentSize);
+            public static readonly int OffsetMask = CalculateOffsetMask(SegmentSize);
+        }
+
+        private static class ValueTypeSegmentHelper<T>
+        {
+            public static readonly int SegmentSize = CalculateSegmentSize(Unsafe.SizeOf<T>());
+            public static readonly int SegmentShift = CalculateSegmentShift(SegmentSize);
+            public static readonly int OffsetMask = CalculateOffsetMask(SegmentSize);
         }
     }
 }