--- a/src/Dependencies/Collections/SegmentedArray`1.cs
+++ b/src/Dependencies/Collections/SegmentedArray`1.cs
@@ -30,17 +30,38 @@ namespace Microsoft.CodeAnalysis.Collections
         /// including any padding the implementation chooses to add. Specifically, array elements lie <c>sizeof</c>
         /// bytes apart.</para>
         /// </remarks>
-        private static readonly int s_segmentSize = SegmentedArrayHelper.CalculateSegmentSize(Unsafe.SizeOf<T>());
+        private static int SegmentSize
+        {
+            [MethodImpl(SegmentedArrayHelper.FastPathMethodImplOptions)]
+            get
+            {
+                return SegmentedArrayHelper.GetSegmentSize<T>();
+            }
+        }
 
         /// <summary>
         /// The bit shift to apply to an array index to get the page index within <see cref="_items"/>.
         /// </summary>
-        private static readonly int s_segmentShift = SegmentedArrayHelper.CalculateSegmentShift(s_segmentSize);
+        private static int SegmentShift
+        {
+            [MethodImpl(SegmentedArrayHelper.FastPathMethodImplOptions)]
+            get
+            {
+                return SegmentedArrayHelper.GetSegmentShift<T>();
+            }
+        }
 
         /// <summary>
         /// The bit mask to apply to an array index to get the index within a page of <see cref="_items"/>.
         /// </summary>
-        private static readonly int s_offsetMask = SegmentedArrayHelper.CalculateOffsetMask(s_segmentSize);
+        private static int OffsetMask
+        {
+            [MethodImpl(SegmentedArrayHelper.FastPathMethodImplOptions)]
+            get
+            {
+                return SegmentedArrayHelper.GetOffsetMask<T>();
+            }
+        }
 
         private readonly int _length;
         private readonly T[][] _items;
@@ -59,17 +80,17 @@ namespace Microsoft.CodeAnalysis.Collections
             }
             else
             {
-                _items = new T[(length + s_segmentSize - 1) >> s_segmentShift][];
+                _items = new T[(length + SegmentSize - 1) >> SegmentShift][];
                 for (var i = 0; i < _items.Length - 1; i++)
                 {
-                    _items[i] = new T[s_segmentSize];
+                    _items[i] = new T[SegmentSize];
                 }
 
                 // Make sure the last page only contains the number of elements required for the desired length. This
                 // collection is not resizeable so any additional padding would be a waste of space.
                 //
                 // Avoid using (length & s_offsetMask) because it doesn't handle a last page size of s_segmentSize.
-                var lastPageSize = length - ((_items.Length - 1) << s_segmentShift);
+                var lastPageSize = length - ((_items.Length - 1) << SegmentShift);
 
                 _items[_items.Length - 1] = new T[lastPageSize];
                 _length = length;
@@ -94,9 +115,10 @@ namespace Microsoft.CodeAnalysis.Collections
 
         public ref T this[int index]
         {
+            [MethodImpl(SegmentedArrayHelper.FastPathMethodImplOptions)]
             get
             {
-                return ref _items[index >> s_segmentShift][index & s_offsetMask];
+                return ref _items[index >> SegmentShift][index & OffsetMask];
             }
         }
 
@@ -135,7 +157,7 @@ namespace Microsoft.CodeAnalysis.Collections
         {
             for (var i = 0; i < _items.Length; i++)
             {
-                _items[i].CopyTo(array, index + (i * s_segmentSize));
+                _items[i].CopyTo(array, index + (i * SegmentSize));
             }
         }
 
@@ -144,7 +166,7 @@ namespace Microsoft.CodeAnalysis.Collections
             for (var i = 0; i < _items.Length; i++)
             {
                 ICollection<T> collection = _items[i];
-                collection.CopyTo(array, arrayIndex + (i * s_segmentSize));
+                collection.CopyTo(array, arrayIndex + (i * SegmentSize));
             }
         }
 
@@ -207,7 +229,7 @@ namespace Microsoft.CodeAnalysis.Collections
                 var index = list.IndexOf(value);
                 if (index >= 0)
                 {
-                    return index + i * s_segmentSize;
+                    return index + i * SegmentSize;
                 }
             }
 
@@ -222,7 +244,7 @@ namespace Microsoft.CodeAnalysis.Collections
                 var index = list.IndexOf(value);
                 if (index >= 0)
                 {
-                    return index + i * s_segmentSize;
+                    return index + i * SegmentSize;
                 }
             }
 
@@ -393,7 +415,7 @@ namespace Microsoft.CodeAnalysis.Collections
                 _array = array;
             }
 
-            public static int SegmentSize => s_segmentSize;
+            public static int SegmentSize => SegmentedArray<T>.SegmentSize;
 
             public T[][] Items => _array._items;
         }