--- a/src/Dependencies/Collections/SegmentedList`1.cs
+++ b/src/Dependencies/Collections/SegmentedList`1.cs
@@ -1049,7 +1049,11 @@ namespace Microsoft.CodeAnalysis.Collections
 
             if (_size > 1)
             {
+#if NET
+                _items.AsSpan(0, _size).Sort(comparison);
+#else
                 SegmentedArray.Sort<T>(_items, 0, _size, Comparer<T>.Create(comparison));
+#endif
             }
             _version++;
         }