* Closes the gap for using `SegmentedArray<int>` (and other unmanaged types) from 2.69x the cost of `int[]` to 1.29x on .NET Framework
* Reduces the penalty for using `SegmentedArray<object>` (and other reference types), particularly on .NET Framework

## Before this change

### net472

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4300.0), X64 RyuJIT
  DefaultJob : .NET Framework 4.8 (4.8.4300.0), X64 RyuJIT


```
|                 Method |  Count |        Mean |    Error |   StdDev | Ratio | RatioSD | Code Size |
|----------------------- |------- |------------:|---------:|---------:|------:|--------:|----------:|
|                &#39;int[]&#39; | 100000 |    78.29 μs | 0.133 μs | 0.104 μs |  1.00 |    0.00 |     107 B |
|             &#39;object[]&#39; | 100000 |   180.58 μs | 0.285 μs | 0.238 μs |  2.31 |    0.00 |     113 B |
|    SegmentedArray&lt;int&gt; | 100000 |   358.65 μs | 1.645 μs | 1.458 μs |  4.58 |    0.02 |     144 B |
| SegmentedArray&lt;object&gt; | 100000 | 1,381.35 μs | 7.889 μs | 7.379 μs | 17.62 |    0.08 |     224 B |

### netcoreapp3.1

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
.NET Core SDK=5.0.103
  [Host]     : .NET Core 3.1.12 (CoreCLR 4.700.21.6504, CoreFX 4.700.21.6905), X64 RyuJIT
  DefaultJob : .NET Core 3.1.12 (CoreCLR 4.700.21.6504, CoreFX 4.700.21.6905), X64 RyuJIT


```
|                 Method |  Count |     Mean |   Error |  StdDev | Ratio | RatioSD | Code Size |
|----------------------- |------- |---------:|--------:|--------:|------:|--------:|----------:|
|                &#39;int[]&#39; | 100000 | 133.1 μs | 0.75 μs | 0.70 μs |  1.00 |    0.00 |     107 B |
|             &#39;object[]&#39; | 100000 | 155.7 μs | 0.66 μs | 0.62 μs |  1.17 |    0.01 |     113 B |
|    SegmentedArray&lt;int&gt; | 100000 | 221.8 μs | 0.74 μs | 0.69 μs |  1.67 |    0.01 |     219 B |
| SegmentedArray&lt;object&gt; | 100000 | 554.6 μs | 2.43 μs | 2.15 μs |  4.16 |    0.03 |     275 B |

### net5.0

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
.NET Core SDK=5.0.103
  [Host]     : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT
  DefaultJob : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT


```
|                 Method |  Count |     Mean |   Error |  StdDev | Ratio | RatioSD | Code Size |
|----------------------- |------- |---------:|--------:|--------:|------:|--------:|----------:|
|                &#39;int[]&#39; | 100000 | 133.2 μs | 0.62 μs | 0.58 μs |  1.00 |    0.00 |     107 B |
|             &#39;object[]&#39; | 100000 | 155.7 μs | 0.77 μs | 0.65 μs |  1.17 |    0.01 |     113 B |
|    SegmentedArray&lt;int&gt; | 100000 | 202.8 μs | 0.46 μs | 0.41 μs |  1.52 |    0.01 |     213 B |
| SegmentedArray&lt;object&gt; | 100000 | 448.2 μs | 1.89 μs | 1.68 μs |  3.37 |    0.02 |     279 B |

## After this change

### net472

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4300.0), X64 RyuJIT
  DefaultJob : .NET Framework 4.8 (4.8.4300.0), X64 RyuJIT


```
|                 Method |  Count |     Mean |   Error |  StdDev | Ratio | Code Size |
|----------------------- |------- |---------:|--------:|--------:|------:|----------:|
|                &#39;int[]&#39; | 100000 | 133.9 μs | 0.25 μs | 0.21 μs |  1.00 |     107 B |
|             &#39;object[]&#39; | 100000 | 157.3 μs | 0.82 μs | 0.73 μs |  1.18 |     113 B |
|    SegmentedArray&lt;int&gt; | 100000 | 172.6 μs | 0.65 μs | 0.61 μs |  1.29 |     205 B |
| SegmentedArray&lt;object&gt; | 100000 | 531.9 μs | 1.88 μs | 1.66 μs |  3.97 |     233 B |

### netcoreapp3.1

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
.NET Core SDK=5.0.103
  [Host]     : .NET Core 3.1.12 (CoreCLR 4.700.21.6504, CoreFX 4.700.21.6905), X64 RyuJIT
  DefaultJob : .NET Core 3.1.12 (CoreCLR 4.700.21.6504, CoreFX 4.700.21.6905), X64 RyuJIT


```
|                 Method |  Count |     Mean |   Error |  StdDev | Ratio | RatioSD | Code Size |
|----------------------- |------- |---------:|--------:|--------:|------:|--------:|----------:|
|                &#39;int[]&#39; | 100000 | 133.2 μs | 0.65 μs | 0.61 μs |  1.00 |    0.00 |     107 B |
|             &#39;object[]&#39; | 100000 | 155.7 μs | 1.04 μs | 0.97 μs |  1.17 |    0.01 |     113 B |
|    SegmentedArray&lt;int&gt; | 100000 | 172.3 μs | 0.59 μs | 0.55 μs |  1.29 |    0.01 |     205 B |
| SegmentedArray&lt;object&gt; | 100000 | 486.8 μs | 1.55 μs | 1.37 μs |  3.66 |    0.02 |     233 B |


### net5.0

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
.NET Core SDK=5.0.103
  [Host]     : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT
  DefaultJob : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT


```
|                 Method |  Count |      Mean |    Error |   StdDev | Ratio | RatioSD | Code Size |
|----------------------- |------- |----------:|---------:|---------:|------:|--------:|----------:|
|                &#39;int[]&#39; | 100000 |  77.69 μs | 0.292 μs | 0.273 μs |  1.00 |    0.00 |     107 B |
|             &#39;object[]&#39; | 100000 | 155.04 μs | 0.427 μs | 0.378 μs |  2.00 |    0.01 |     113 B |
|    SegmentedArray&lt;int&gt; | 100000 | 166.89 μs | 0.550 μs | 0.515 μs |  2.15 |    0.01 |     203 B |
| SegmentedArray&lt;object&gt; | 100000 | 434.87 μs | 1.560 μs | 1.459 μs |  5.60 |    0.03 |     232 B |