@sharwell 2021-02-21
I was able to further optimize the net5.0 case by avoiding array variance checks.

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19042
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
.NET Core SDK=5.0.103
  [Host]     : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT
  DefaultJob : .NET Core 5.0.3 (CoreCLR 5.0.321.7212, CoreFX 5.0.321.7212), X64 RyuJIT


```
|                 Method |  Count |      Mean |    Error |   StdDev | Ratio | Code Size |
|----------------------- |------- |----------:|---------:|---------:|------:|----------:|
|                &#39;int[]&#39; | 100000 |  77.67 μs | 0.232 μs | 0.206 μs |  1.00 |     107 B |
|             &#39;object[]&#39; | 100000 | 155.54 μs | 0.740 μs | 0.692 μs |  2.00 |     113 B |
|    SegmentedArray&lt;int&gt; | 100000 | 172.87 μs | 0.747 μs | 0.699 μs |  2.23 |     208 B |
| SegmentedArray&lt;object&gt; | 100000 | 310.59 μs | 1.195 μs | 1.118 μs |  4.00 |     221 B |

The numbers were actually a bit better for `SegmentedArray<object>` before adding the `IsValueType` check, but they were much worse for `SegmentedArray<int>`.